html and pdf render to generate class
